\name{SummaryCovNA-class}
\docType{class}
\alias{SummaryCovNA-class}
\alias{show,SummaryCovNA-method}

\title{Class "SummaryCovNA" - summary of "CovNA" objects  }

\description{The "CovNA" object plus some additional summary information}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("SummaryCovNA", ...)},
but most often by invoking 'summary' on a "CovNA" object. They contain values 
meant for printing by 'show'.
}
\section{Slots}{
%    \describe{  }
No Slots defined with class \code{"SummaryCovNA"} in the signature.
}
\section{Methods}{
%  \describe{     }
No Methods defined with class \code{"SummaryCovNA"} in the signature.
}
\references{
  Todorov V & Filzmoser P (2009),
  An Object Oriented Framework for Robust Multivariate Analysis.
  \emph{Journal of Statistical Software}, \bold{32}(3), 1--47.
  <doi:10.18637/jss.v032.i03>.
}
\author{Valentin Todorov \email{valentin.todorov@chello.at}}
%\note{}
\seealso{
  \code{\link{CovNA-class}}
}
\examples{
showClass("SummaryCovNA")
}
\keyword{classes}
